Remove splash destination, add deep links and fix bottom navigation
Adds a deep link to the speakers screen and to the sessions screen.
Adds the "confetti/confetti" scheme to the manifest just like the wear
app does too.

Fix #551 by making bottom nav destinations save their state and restore
it when returning back to them. Also ensures that when navigating
between them, the entire backstack is not popped, but the start
destination stays in the backstack, just like NiA handles this

Super important to note here that:
We seem to have to clear the backstack of the bottom nav top level
destinations when we go away from that flow and go into the conferences
screen. If we do not do that, the previously saved destinations that
we are saving the state of as we pop them off will take precedence
over the new routes we pass in. This means that they will be
referencing the old conference argument, despite us sending in a new
one, which seems to be completely ignored. So if we try to go to
"speakers/kotlinconf2023" but we've before that popped a
"spekers/fosdem2023" destination and saved the state, it will ignore
our "kotlinconf2023" parameter and still have "fosdem2023" in its
backstack parameters.
We do this inside `onSwitchConference` for now

Adds official splash screen support and add a SplashReadyStatus class
to hold the state. The top level destinations can use this to report
when they are ready to get rid of it. In the future this can be
postponed even further to make sure that the sessions screen waits not
only until it is ready to load the data, but after it has already
started displaying the first batch of data.
For now the first two screens that one can see when opening the app are
either the conferences destinations or the sessions destinations, and
both of these report when they are ready to show. If we are opening an
app using a deep link, we simply do not postpone getting rid of the
splash screen in the first place, and let those screens show
immediately.