Recreate the iOS app UI on a new deep link incoming
With this change whenever a deep link is received, we create a
completely new AppComponent so that we start from a clean slate.
In order to be able to observe this change from SwiftUI we also store a
local @State object which we increment at the same time in order to
trigger a re-render of the View.
This is absolutely a hack and we should find a better approach as soon
as possible